django_app/requirements.txt
Django>=4.2,<6
crewai>=0.86,<2
python-dotenv>=1.0,<2
fpdf2>=2.7,<3
